refactor: create crate dedicated to testing
