+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["lib", "bin"]
+members = ["lib", "bin", "testing"]
 resolver = "2"
