+++ bin/src/lib.rs
@@ -1,3 +1,2 @@
 pub mod application;
 pub mod commands;
-pub mod testing;
