--- bin/src/testing/mod.rs
@@ -1,2 +1,0 @@
-pub mod arrangements;
-pub mod assertions;
