+++ bin/tests/change_command_speakers_only_should.rs
@@ -1,12 +1,13 @@
-use convertible_couch::{
-    application::{ApplicationChangeResult, ApplicationResult},
-    testing::arrangements::{ApplicationBuilder, ArgumentsBuilder},
-};
+use convertible_couch::application::{ApplicationChangeResult, ApplicationResult};
 use convertible_couch_lib::{
-    application_error::ApplicationError,
+    application_error::ApplicationError, speakers_settings::SpeakersSettingsResult,
+};
+use convertible_couch_testing::{
+    arrangements::{
+        builders::{ApplicationBuilder, ArgumentsBuilder},
+        fuzzing::{ComputerBuilder, Fuzzer},
+    },
     func,
-    speakers_settings::SpeakersSettingsResult,
-    testing::fuzzing::{ComputerBuilder, Fuzzer},
 };
 
 #[test]
