+++ lib/Cargo.toml
@@ -11,7 +11,6 @@ license-file = "../LICENSE"
 [dependencies]
 log = "0.4.29"
 log4rs = "1.4.0"
-rand = "0.10.0"
 thiserror = "2.0.18"
 
 [target.'cfg(windows)'.dependencies]
@@ -19,4 +18,5 @@ windows = { version = "0.62.2", features = ["Win32_Devices_Display", "Win32_Devi
 windows-core = "0.62.2"
 
 [dev-dependencies]
+convertible_couch_testing = { path = "../testing" }
 test-case = "3.3.1"
