+++ lib/src/lib.rs
@@ -2,4 +2,3 @@ pub mod application_error;
 pub mod displays_settings;
 pub mod log;
 pub mod speakers_settings;
-pub mod testing;
