--- lib/src/testing/mod.rs
@@ -1,2 +1,0 @@
-pub mod fuzzing;
-pub mod utils;
