+++ testing/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "convertible_couch_testing"
+version = "2.0.1"
+edition = "2024"
+authors = ["Thomas PEREZ"]
+description = "A library for helping to write tests for convertible_couch."
+readme = "../README.md"
+repository = "https://github.com/drsanguin/convertible_couch"
+license-file = "../LICENSE"
+
+[dependencies]
+convertible_couch = { path = "../bin" }
+convertible_couch_lib = { path = "../lib" }
+rand = "0.10.0"
+windows = "0.62.2"
+windows-core = "0.62.2"
