+++ testing/src/arrangements/fuzzing/displays/display_name.rs
