+++ testing/src/arrangements/fuzzing/displays/gsm_id.rs
