+++ testing/src/arrangements/fuzzing/displays/resolution.rs
