+++ testing/src/arrangements/fuzzing/speakers/settings_api/mod.rs
@@ -1,4 +1,4 @@
-use crate::testing::fuzzing::speakers::{
+use crate::arrangements::fuzzing::speakers::{
     FuzzedSpeaker, settings_api::behaviour::CurrentFuzzedSpeakersSettingsApiBehaviour,
 };
