+++ testing/src/arrangements/fuzzing/speakers/speaker_name.rs
